Struct bookworm_records::book::Book

pub struct Book {
    pub id: Option<BookID>,
    pub title: Option<String>,
    pub authors: Option<Vec<String>>,
    pub series: Option<Series>,
    pub description: Option<String>,
    pub variants: Vec<BookVariant>,
    pub free_tags: HashSet<String>,
    pub named_tags: HashMap<String, String>,
}

Stores the metadata for a specific book, with an associated ID, variants, and fields shared between different variants.

Variants can be different file formats of the same book, or different realizations of the same book.

Fields
id: Option<BookID>title: Option<String>authors: Option<Vec<String>>series: Option<Series>description: Option<String>variants: Vec<BookVariant>free_tags: HashSet<String>named_tags: HashMap<String, String>
Implementations

impl Book

pub fn authors(&self) -> Option<&[String]>

pub fn title(&self) -> Option<&str>

pub fn series(&self) -> Option<&Series>

pub fn variants(&self) -> &[BookVariant]

pub fn tags(&self) -> &HashMap<String, String>

pub fn free_tags(&self) -> &HashSet<String>

pub fn description(&self) -> Option<&String>

pub fn get_column(&self, column: &ColumnIdentifier) -> Option<Cow<'_, str>>

pub fn push_variant(&mut self, variant: BookVariant)

pub fn merge_mut(&mut self, other: &Self)

Merges self with other, assuming that other is in fact a variant of self. Missing metadata will be utilized from other, and self variants will be extended to include other variants.

Arguments

other - Another book with additional metadata, to be merged with self

impl Book

pub fn placeholder() -> Book

Generates a placeholder book, which should not be used except as a way to reduce the run-time of specific operations. Placeholders do not have an ID, and calling id() on a placeholder will result in a panic.

pub fn is_placeholder(&self) -> bool

Returns true if self is a placeholder

pub fn from_variant(id: BookID, variant: BookVariant) -> Book

Creates a Book with the given ID and core metadata.

pub fn id(&self) -> BookID

Returns the ID, which can be used as an index.

Errors

Will panic if self was created by placeholder.

impl Book

pub fn set_column<S: AsRef<str>>(
&mut self,
column: &ColumnIdentifier,
value: S
) -> Result<(), RecordError>

Sets specified columns, to the specified value. Titles will be stored directly, authors will be stored as a list containing a single author. ID and Variants can not be modified through set_column. Series will be parsed to extract an index - strings in the form “series [num]” will be parsed as (“series”, num).

All remaining arguments will be stored literally.

Arguments

column - the column of interest.

value - the value to store into the column.

pub fn edit_column<E: AsRef<Edit>>(
&mut self,
column: &ColumnIdentifier,
edit: E
) -> Result<(), RecordError>

pub fn extend_column<S: AsRef<str>>(
&mut self,
column: &ColumnIdentifier,
value: S
) -> Result<(), RecordError>

pub fn delete_column(
&mut self,
column: &ColumnIdentifier
) -> Result<(), RecordError>

pub fn cmp_column(&self, other: &Self, column: &ColumnIdentifier) -> Ordering

Compares the specified columns, by default using Rust’s native string comparison

missing fields will be compared as empty strings. If ID is compared, numerical comparison will occur. If Series is compared, the series will be compared first, with the index being used as a tie-breaker.

Arguments

other - The other book with column to compare.

column - the column of interest.

pub fn cmp_columns(
&self,
other: &Self,
columns: &[(ColumnIdentifier, ColumnOrder)]
) -> Ordering